moved guest login into auth route, updated fe api calls accordingly and also changed testing to automatically invalidate redis key over using time.sleep
