+++ backend/server/routers/auth.py
@@ -2,7 +2,8 @@
 from typing import Annotated, Dict, Optional, Tuple, Union, cast
 from secrets import token_hex, token_urlsafe
 from time import time
-from fastapi import APIRouter, Cookie, HTTPException, Response, Security
+from fastapi import APIRouter, Cookie, Depends, HTTPException, Response, Security
+from fastapi_limiter.depends import RateLimiter
 from pydantic import BaseModel
 from starlette.status import HTTP_401_UNAUTHORIZED, HTTP_400_BAD_REQUEST, HTTP_500_INTERNAL_SERVER_ERROR
 
@@ -11,7 +12,7 @@ from server.db.helpers.models import NotSetupUserStorage, GuestSessionInfoModel,
 from server.db.helpers.users import delete_user, insert_new_user
 
 from .utility.sessions.errors import ExpiredRefreshTokenError, ExpiredSessionTokenError, OldRefreshTokenError
-from .utility.sessions.interface import create_new_guest_token_pair, get_session_info_from_refresh_token, get_session_info_from_session_token, logout_session, setup_new_devsoc_session, create_new_devsoc_token_pair
+from .utility.sessions.interface import create_new_guest_token_pair, get_session_info_from_refresh_token, get_session_info_from_session_token, logout_session, setup_new_devsoc_session, setup_new_guest_session, create_new_devsoc_token_pair
 
 from .utility.sessions.middleware import HTTPBearer401, set_secure_cookie
 from .utility.oidc.requests import DecodedIDToken, exchange_and_validate, generate_oidc_auth_url, get_userinfo_and_validate, refresh_and_validate, revoke_token, validate_authorization_response
@@ -212,6 +213,22 @@ def login(res: Response, payload: ExchangeCodePayload, next_auth_state: Annotate
 
     # set the cookies and return the identity
     set_secure_cookie(res, AUTH_STATE_COOKIE, None)
+    set_secure_cookie(res, REFRESH_TOKEN_COOKIE, new_refresh_token, refresh_expiry)
+    return IdentityPayload(session_token=new_session_token, exp=session_expiry, uid=uid)
+
+
+
+@router.post('/guest_login',
+             dependencies=[Depends(RateLimiter(times=3, seconds=60))],
+            )
+def create_guest_session(res: Response) -> IdentityPayload:
+    # create new login session for user in db, generating new tokens
+    uid = insert_new_guest_user()
+    new_session_token, session_expiry, new_refresh_token, refresh_expiry = setup_new_guest_session(uid)
+
+    # TODO-OLLI(pm): setting up proper logging
+
+    # set the cookies and return the identity
     set_secure_cookie(res, REFRESH_TOKEN_COOKIE, new_refresh_token, refresh_expiry)
     return IdentityPayload(session_token=new_session_token, exp=session_expiry, uid=uid)
