+++ backend/server/tests/user/test_guest_login_rate_limit.py
@@ -1,6 +1,7 @@
 import os
 import time
 
+import redis
 import requests
 from dotenv import load_dotenv
 
@@ -8,12 +9,32 @@ load_dotenv("../env/backend.env")
 os.environ["SESSIONSDB_SERVICE_HOSTNAME"] = "localhost"
 os.environ["MONGODB_SERVICE_HOSTNAME"] = "localhost"
 
+from server.db.redis.limiter_conn import limiter_redis_kwargs
 from server.db.redis.setup import reset_redis_limiterdb
 
-GUEST_LOGIN_URL = "http://127.0.0.1:8000/dev/guest_login"
+GUEST_LOGIN_URL = "http://127.0.0.1:8000/auth/guest_login"
 
 RATE_LIMIT_TIMES = 3
-RATE_LIMIT_SECONDS = 60
+
+def fast_forward_limiter_window():
+    """Fast-forward past the rate-limit window without sleeping for its full
+    duration.
+
+    The window is enforced by a Redis TTL that fastapi-limiter sets server-side,
+    so a Python time-machine (freezegun/time-machine) in this test process can't
+    advance it. Expiring the counter keys directly in Redis is the equivalent
+    'time machine' here, and doubles as a check that a TTL was actually set.
+    """
+    limiter_db = redis.Redis(**limiter_redis_kwargs())
+    try:
+        keys = limiter_db.keys("*")
+        assert keys, "expected limiter counters to exist before the window expires"
+        for key in keys:
+            assert limiter_db.pttl(key) > 0, "limiter counter should carry a TTL"
+            limiter_db.pexpire(key, 1)  # expire ~immediately
+    finally:
+        limiter_db.close()
+    time.sleep(0.05)  # let redis apply the expiry
 
 def test_guest_login_allows_requests_up_to_limit():
     reset_redis_limiterdb()
@@ -40,7 +61,7 @@ def test_guest_login_allows_requests_again_after_window_expires():
         requests.post(GUEST_LOGIN_URL, timeout=5000)
     assert requests.post(GUEST_LOGIN_URL, timeout=5000).status_code == 429
 
-    time.sleep(RATE_LIMIT_SECONDS + 2)
+    fast_forward_limiter_window()
 
     res = requests.post(GUEST_LOGIN_URL, timeout=5000)
     assert res.status_code == 200
